An S4 class to represent a PWM matrix.

An object of class 'pwm' represents the alphabet*width position weight matrix of a sequence motif. In case of DNA sequence motif, the entry in row i, column j gives the probability of observing nucleotide c('A','C','G','T')[i] in position j of the motif.

# S4 method for pwm
show(object)

# S4 method for pwm
summary(object, ...)

# S4 method for pwm,ANY
plot(x, y = "missing", ...)

# S4 method for pwm
pwm(pwm)

# S4 method for pwm
ic(pwm)

# S4 method for pwm
consensus(pwm)

Arguments

object
object of pwm-class

...
additional parameters for plot function

x
object of pwm-class

y
default (missing) for plot function

pwm
object of pwm-class

Value

pwm-class object with slots: pwm, width, ic and alphabet.

Functions

show,pwm-method: Shows the position weight matrix.

summary,pwm-method: Prints the summary information about position weight matrix.

plot,pwm,ANY-method: Plots the sequence logo of the position weight matrix.

pwm,pwm-method: Access to 'pwm' slot

ic,pwm-method: Access to 'ic' slot

consensus,pwm-method: Access to 'consensus' slot

Slots

pwm

matrix. The position weight matrix.

width

numeric. The width of the motif.

ic

numeric. The information content (IC).

alphabet

character. The sequence alphabet. Currently, only 'DNA' and 'RNA' is supported.

consensus

character. The consensus sequence.

Examples

mFile <- system.file("extdata/pwm1", package = "seqLogo")
m <- read.table(mFile)
p <- makePWM(m)
#
# slot access
pwm(p)

#>     1   2   3   4   5   6   7   8
#> A 0.0 0.0 0.0 0.3 0.2 0.0 0.0 0.0
#> C 0.8 0.2 0.8 0.3 0.4 0.2 0.8 0.2
#> G 0.2 0.8 0.2 0.4 0.3 0.8 0.2 0.8
#> T 0.0 0.0 0.0 0.0 0.1 0.0 0.0 0.0

ic(p)

#> [1] 1.2780719 1.2780719 1.2780719 0.4290494 0.1535607 1.2780719 1.2780719
#> [8] 1.2780719

consensus(p)

#> [1] "CGCGCGCG"